requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
pydantic
pydantic-settings
python-dotenv
loguru

#PDFPARSER
scikit-learn
Pillow
transformers
numpy
python-dotenv
torch
tqdm
tabulate
ftfy
texify
rapidfuzz
surya-ocr
filetype
regex
pdftext
grpcio
python-multipart
pypdfium2
